Parsing arguments exposed as a separate function
